Move LogCat prop to SentryOptions
