+++ src/Sentry/Platforms/Android/SentryOptions.cs
@@ -198,20 +198,6 @@ public partial class SentryOptions
         /// </summary>
         public TimeSpan ReadTimeout { get; set; } = TimeSpan.FromSeconds(5);
 
-        /// <summary>
-        /// Gets or sets whether when LogCat logs are attached to events.
-        /// The default is <see cref="LogCatIntegrationType.None"/>
-        /// </summary>
-        /// <seealso cref="LogCatMaxLines" />
-        public LogCatIntegrationType LogCatIntegration { get; set; } = LogCatIntegrationType.None;
-
-        /// <summary>
-        /// Gets or sets the maximum number of lines to read from LogCat logs.
-        /// The default value is 1000.
-        /// </summary>
-        /// <seealso cref="LogCatIntegration" />
-        public int LogCatMaxLines { get; set; } = 1000;
-
 
         // ---------- Other ----------
